requirements.txt
numpy
pandas
python-dotenv
pymongo
pymongo[srv]==3.11
certifi
scikit-learn
streamlit
pyyaml
